Rewrite to be mostly webNavigation-based
With webNavigation events, it's possible to detect when the user navigates to the new tab page and what type of navigation it is (specifically, forward_back).

Because Chrome doesn't provide button info in mouseup events, the content script keeps track of whether there was just a back button down event before it takes action on a button up event. This prevents things like closing a tab on button down and the remaining tab getting the button up.

The content script uses the history size to determine whether to close the tab or let the back press go through and trigger navigation events.